+++ src/lib/package/manifest.rs
@@ -336,7 +336,7 @@ license = 'MIT'
 
 [dependencies]
 'awesome/a' = '>= 1.0.0 < 2.0.0'
-'cool/b' = { git = 'https://github.com/super/cool', branch = "v1.0.0" }
+'cool/b' = { git = 'https://github.com/super/cool', tag = "v1.0.0" }
 'great/c' = { path = 'here/right/now' }
 
 [dev_dependencies]